requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httplib2==0.22.0
idna==3.8
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.3
